+++ WebApiRest/Data/UsuarioData.cs
@@ -49,7 +49,7 @@ namespace WebApiRest.Data
                         Ciudad = dr["ciudad"].ToString(),                        
                         IdArea = dr["idArea"].ToString(),
                         Area = dr["area"].ToString(),
-                        Estado = Convert.ToInt16(dr["estado"].ToString()),
+                        Estado = Convert.ToInt32(dr["estado"].ToString()),
                         FechaCreacion = Convert.ToDateTime(dr["fechaCreacion"].ToString()),
                         FechaModificacion = Convert.ToDateTime(dr["fechaModificacion"].ToString()),
                     };
@@ -57,7 +57,7 @@ namespace WebApiRest.Data
                 await dr.NextResultAsync();
 
                 item.Info = cmd.Parameters["@info"].Value.ToString();
-                item.Error = Convert.ToInt16(cmd.Parameters["@error"].Value.ToString());
+                item.Error = Convert.ToInt32(cmd.Parameters["@error"].Value.ToString());
             }
             catch (Exception ex)
             {
@@ -109,7 +109,7 @@ namespace WebApiRest.Data
                         Ciudad = dr["ciudad"].ToString(),                        
                         IdArea = dr["idArea"].ToString(),
                         Area = dr["area"].ToString(),                        
-                        Estado = Convert.ToInt16(dr["estado"].ToString()),
+                        Estado = Convert.ToInt32(dr["estado"].ToString()),
                         FechaCreacion = Convert.ToDateTime(dr["fechaCreacion"].ToString()),
                         FechaModificacion = Convert.ToDateTime(dr["fechaModificacion"].ToString()),
                     };
@@ -117,7 +117,7 @@ namespace WebApiRest.Data
                 await dr.NextResultAsync();
 
                 item.Info = cmd.Parameters["@info"].Value.ToString();
-                item.Error = Convert.ToInt16(cmd.Parameters["@error"].Value.ToString());
+                item.Error = Convert.ToInt32(cmd.Parameters["@error"].Value.ToString());
             }
             catch (Exception ex)
             {
@@ -163,7 +163,7 @@ namespace WebApiRest.Data
                 await cmd.ExecuteNonQueryAsync();                
 
                 response.Info = cmd.Parameters["@info"].Value.ToString();
-                response.Error = Convert.ToInt16(cmd.Parameters["@error"].Value.ToString());
+                response.Error = Convert.ToInt32(cmd.Parameters["@error"].Value.ToString());
 
             }
             catch (Exception ex)
@@ -218,7 +218,7 @@ namespace WebApiRest.Data
                 await cmd.ExecuteNonQueryAsync();                
 
                 response.Info = cmd.Parameters["@info"].Value.ToString();
-                response.Error = Convert.ToInt16(cmd.Parameters["@error"].Value.ToString());
+                response.Error = Convert.ToInt32(cmd.Parameters["@error"].Value.ToString());
 
             }
             catch (Exception ex)
